Add options for modifying the global controlDict values/switches from the command-line
The following command-line options are available:
 -DebugSwitches        key=value
 -DimensionedConstants key=value
 -InfoSwitches         key=value
 -OptimisationSwitches key=value
 -Tolerances           key=value

In order to modify more than one key/value pair for a given category, you will
need to specify a given option more than once.

If an invalid/unknown key is specified, a complete list of valid keys for the
application will be written at the console.

Martin Beaudoin 05/2014