testHarness: adding the environment variable CDASH_SCM_INFO in order to override the SCM information that gets appended at the end of the buildname string.
The buildname string is the string that will identify your test run when
published on the foam-extend CDash service.

When using git, this SCM information is picked up automatically from the CTest
script: -git-branch=the_git_branch_name-git-rev=the_git_revision_number.

But when using a non-git SCM (mercurial, etc), when the foam-extend source
code you are testing is not under a valid git repository, or simply when a user
needs to add more details to the standard SCM information (branch name and
revision number), then setting the environment variable CDASH_SCM_INFO to the
proper information will allow you to override the suffix we add at the end of
the buildname.

Please note that in order to keep the CMake script simple, for non-git
repositories, the fallback solution to try grabbing the information through
Mercurial has been removed. Please use the CDASH_SCM_INFO environment
varible instead.