Correction for a problem with newer versions of CMake that are refusing to upload the test harness results  to the CDash server. The generated error message is the following: "Error message was: Illegal characters found in URL"
Diagnostic:
    If the environment variable CDASH_SUBMIT_LOCAL_HOST_ID is not initialized, the FQN hostname of the
system will be picked by the CMake/CTest scripts. The CTest internal system command used for retrieving
the FQN hostname is indeed leaving a carriage return in the returned hostname string.
That string is in turn reused to construct an URL that will be shared with the CDash service. The presence
of such characters (\n and \r) is no longer accepted by newer versions of CTest/CMake.

Solution:
    The solution is simply to check and remove such characters from the hostname string.