Bugfix in MixingPlaneInterpolation due to recent GGI updates
MixingPlaneInterpolation is actually implemented in terms of GGIIinterpolation
and that one now has handling of partially overlapping faces (needed for bridge
overlap option). There's a tolerance GGIUncoveredFaceAreaTol (defaults to 0.999)
which needs to be set to 0 in case of Mixing Plane to properly re-scale the
weights. This is now done in MixingPlaneInterpolation.